backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0
anthropic==0.28.0
openai-whisper==20231117
librosa==0.10.2
praat-parselmouth==0.4.3
numpy==1.26.4
soundfile==0.12.1
scipy==1.13.0
twilio==9.2.3
reportlab==4.2.2
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.7.1
aiofiles==23.2.1
transformers==4.46.3
torch==2.4.1
torchaudio==2.4.1
scikit-learn==1.5.2
joblib==1.4.2
kaggle==1.6.17
imageio-ffmpeg==0.5.1
pydub==0.25.1
opensmile==2.5.0
speechbrain==1.0.0
xgboost==2.0.3

# Twilio Media Streams + Audio Processing
webrtcvad==2.0.10
audioop-lts==0.2.1

# Celery Task Queue
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# AWS S3
boto3==1.34.69

# Speaker Diarization
pyannote.audio==3.1.1

# Email Notifications
sendgrid==6.11.0

# Database (optional - for PostgreSQL in production)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
